# swagger.yaml
swagger: "2.0"
info:
  description: "A microservice called cloudwatch-logs-adapter observing AWS logs from different microservices. It serves a specific amount of last log lines of corresponding microservice identified by name. Ticket at [http://swagger.io](http://swagger.io)."
  version: "1.0.0"
  title: "Cloudwatch Logs Adapter"
  contact:
    email: "david.renner@consort-it.de"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "consort-it.de"
basePath: "/api/v1/cloudwatch-logs-adapter"
schemes:
- "http"
paths:
  /groups:
    get:
      tags:
      - "CloudwatchLogs"
      summary: "Receive List of available LogGroups with metadata."
      description: "This URI serves all available LogGroups in AWS CloudWatch. Each LogGroup contains a set of information except LogStream details."
      operationId: "getGroups"
      produces:
      - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Group"
        400:
          description: "Client sent wrong or missed crucial information and should consider sending a proper request again."
          schema:
            $ref: "#/definitions/ErrorResponse"
        401:
          description: "Access is forbidden using given credentials."
          schema:
            $ref: "#/definitions/ErrorResponse"
        403:
          description: "Access is required but no Authentication information has been given."
          schema:
            $ref: "#/definitions/ErrorResponse"
        500:
          description: "Service did not behave well and made a mistake that caused itself to fail."
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
  Group:
    type: "object"
    required:
    - logGroupName
    - creationTime
    - arn
    - storedBytes
    properties:
      logGroupName:
        type: "string"
        description: "Group identifier or name for finding this Group."
      creationTime:
        type: "string"
        format: "date-time"
        description: "Group Creation Time"
      retentionInDays:
        type: "integer"
        format: "int64"
        description: "Group Age in days"
      metricFilterCount:
        type: "integer"
        format: "int64"
        description: "Group metric filter count"
      arn:
        type: "string"
        description: "Group Address Resolution Name"
      storedBytes:
        type: "integer"
        format: "int64"
        description: "Group Size in Storage"
    xml:
      name: "group"
...
  ErrorResponse:
    type: object
    required:
    - status
    - code
    - message
    - location
    - time
    properties:
      code:
        type: string
        pattern: 'CLA-\d+$'
        description: "This error code should have the format 'CLA-XXX' where XXX is an integer. It is used to uniquely distinguish different error cases in order to display a suitable and translatable error message on the UI."
        example: "CLA-1234"
      status:
        type: integer
        description: "Represents the http status error that goes along with this error."
      message:
        type: string
        description: "Should contain a short, meaningful description of the error case. Might be displayed to the end user."
        example: "Could not create service 'some-name' because this service already exists."
      trace:
        type: array
        description: "Contains a trace of errors if available. Only use for forwarding to developer. In case this is missing as it is optional use location to identify where error happened originally."
        items:
          $ref: "#/definitions/TracePoint"
      location:
        type: string
        description: "Field that indicates where the error occured. This is a mandatory field and should be considered in case trace is not available."
      time:
        type: "string"
        format: "date-time"
        description: "The exact time the error occured within microservice."
  TracePoint:
    type: "object"
    required:
    - methodName
    - fileName
    - lineNumber
    - className
    - nativeMethod
    properties:
      methodName:
        type: "string"
      fileName:
        type: "string"
      lineNumber:
        type: "integer"
      className:
        type: "string"
      nativeMethod:
        type: "boolean"